[refactor]: edit directory name
