+++ app/build.gradle
@@ -5,11 +5,11 @@ plugins {
 }
 
 android {
-    namespace 'org.sopt.dosopttemplate'
+    namespace 'org.sopt.dosoptjaewon'
     compileSdk 33
 
     defaultConfig {
-        applicationId "org.sopt.dosopttemplate"
+        applicationId "org.sopt.dosoptjaewon"
         minSdk 28
         targetSdk 33
         versionCode 1
@@ -25,11 +25,11 @@ android {
         }
     }
     compileOptions {
-        sourceCompatibility JavaVersion.VERSION_11
-        targetCompatibility JavaVersion.VERSION_11
+        sourceCompatibility JavaVersion.VERSION_17
+        targetCompatibility JavaVersion.VERSION_17
     }
     kotlinOptions {
-        jvmTarget = '11'
+        jvmTarget = '17'
     }
     viewBinding{
         enabled = true
@@ -40,7 +40,7 @@ dependencies {
 
     implementation 'androidx.core:core-ktx:1.8.0'
     implementation 'androidx.appcompat:appcompat:1.6.1'
-    implementation 'com.google.android.material:material:1.5.0'
+    implementation 'com.google.android.material:material:1.9.0'
     implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.5'
