+++ app/src/androidTest/java/org/sopt/dosoptjaewon/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package org.sopt.dosopttemplate
+package org.sopt.dosoptjaewon
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
