+++ app/src/main/AndroidManifest.xml
@@ -10,7 +10,7 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.DoSoptTemplate"
+        android:theme="@style/Theme.DoSoptJaewon"
         tools:targetApi="31">
         <activity
             android:name=".presentation.signup.SignupActivity"
