+++ app/src/main/res/values/themes.xml
@@ -1,9 +1,9 @@
-<resources xmlns:tools="http://schemas.android.com/tools">
+<resources>
     <!-- Base application theme. -->
-    <style name="Base.Theme.DoSoptTemplate" parent="Theme.Material3.DayNight.NoActionBar">
+    <style name="Base.Theme.DoSoptJaewon" parent="Theme.Material3.DayNight.NoActionBar">
         <!-- Customize your light theme here. -->
         <!-- <item name="colorPrimary">@color/my_light_primary</item> -->
     </style>
 
-    <style name="Theme.DoSoptTemplate" parent="Base.Theme.DoSoptTemplate" />
+    <style name="Theme.DoSoptJaewon" parent="Base.Theme.DoSoptJaewon" />
 </resources>
