+++ settings.gradle
@@ -12,5 +12,5 @@ dependencyResolutionManagement {
         mavenCentral()
     }
 }
-rootProject.name = "DoSoptTemplate"
+rootProject.name = "DoSoptJaewon"
 include ':app'
